requirements.txt
Flask>=3.0.0
requests>=2.31.0
markdown>=3.5.0
python-frontmatter>=1.1.0
Pillow>=10.0.0
python-magic-bin>=0.4.14
openai>=1.0.0
python-dotenv>=1.0.0
